% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopGridApp.R
\name{PopGridApp}
\alias{PopGridApp}
\title{Shiny app to aggregate block-level population data from the decennial Census to a chosen grid definition}
\usage{
PopGridApp()
}
\value{
\link{PopGridApp} does not return a value but instead saves three output files: a shapefile of the selected grid with population data, a CSV with gridded population for each of the race-ethnicity-gender-age demographic groups included in BenMAP, and a CSV with county weights for each of the eight race-ethnic groups for each grid cell.
}
\description{
This function runs PopGrid via a shiny app UI, offering the same functionality as \link{run_aggregation} but with a simple way to provide inputs.
}
\examples{
PopGridApp()

}
